Modernize python-augeas spec file
Migrate the spec file to use Fedora's modern Python `pyproject` macros.
This leverages automatic buildrequires discovery, and aligns the package
with current Fedora packaging standards.

The %changelog was removed and changed to %autochangelog as it is usable
in Fedora, and does not require the changelog to be maintained in the
upstream repository.

Signed-off-by: Rafael Guterres Jeffman <rjeffman@redhat.com>